// src/api/types/fieldType.d.ts
export type CAREER_RANGE = "0_4" | "4_10" | "10_100";
export type ROLE_MAIN = "pm" | "design" | "dev";
export type ROLE_DETAIL = "product" | "ux_ui" | "app" | "web_front" | "back_end";
export type TEAM_CATEGORY = "portfolio" | "side_project";
export type MEMBER_STATUS = "apply" | "allow" | "deny";
export type TEAM_PARTY_MEMBER_STATUS = MEMBER_STATUS | "expired";